Move more db packages into db2, add logging to AllStatements
